refactor(types): moves types
