+++ src/methods/common.ts
@@ -1,5 +1,5 @@
 import * as ERRORS from "../errors.js";
-import type { Lyra, PropertiesSchema, ResolveSchema } from "../types.js";
+import type { Lyra, PropertiesSchema, ResolveSchema } from "../types/index.js";
 import type { SearchParams } from "./search.js";
 
 export function assertDocSchema<S extends PropertiesSchema>(doc: ResolveSchema<S>, lyraSchema: PropertiesSchema) {
