+++ src/methods/create.ts
@@ -1,4 +1,4 @@
-import type { Configuration, Lyra, PropertiesSchema } from "../types.js";
+import type { Configuration, Lyra, PropertiesSchema } from "../types/index.js";
 import { defaultTokenizerConfig, Language } from "../tokenizer/index.js";
 import * as ERRORS from "../errors.js";
 import { create as createNode } from "../radix-tree/node.js";
