+++ src/methods/remove.ts
@@ -1,5 +1,5 @@
 import type { RadixNode } from "src/radix-tree/node.js";
-import type { Lyra, PropertiesSchema, ResolveSchema } from "../types.js";
+import type { Lyra, PropertiesSchema, ResolveSchema } from "../types/index.js";
 import { defaultTokenizerConfig } from "../tokenizer/index.js";
 import { removeDocumentByWord } from "../radix-tree/radix.js";
 import * as ERRORS from "../errors.js";
