+++ src/methods/search.ts
@@ -1,5 +1,5 @@
 import type { RadixNode } from "src/radix-tree/node.js";
-import type { Lyra, PropertiesSchema, ResolveSchema, SearchProperties, TokenMap, TokenScore, BM25Params, BM25OptionalParams, PropertiesBoost, FacetsSearch } from "../types.js";
+import type { Lyra, PropertiesSchema, ResolveSchema, SearchProperties, TokenMap, TokenScore, BM25Params, BM25OptionalParams, PropertiesBoost, FacetsSearch, WhereFilter } from "../types/index.js";
 import { defaultTokenizerConfig, Language } from "../tokenizer/index.js";
 import { find as radixFind } from "../radix-tree/radix.js";
 import { formatNanoseconds, getNanosecondsTime, sortTokenScorePredicate } from "../utils.js";
@@ -105,6 +105,25 @@ export type SearchParams<S extends PropertiesSchema> = {
    * });
    */
   facets?: FacetsSearch<S>;
+
+  /**
+   * Filter the search results.
+   * 
+   * @example
+   * // Search for documents that contain 'Headphones' in the 'description' field and
+   * // have a price less than 100.
+   * 
+   * const result = await search(db, {
+   *  term: 'Headphones',
+   *  properties: ['description'],
+   *  where: {
+   *    price: {
+   *      lt: 100
+   *    }
+   *  }
+   * });
+   */
+  where?: WhereFilter<S>;
 };
 
 export type SearchResult<S extends PropertiesSchema> = {
