+++ src/types/index.ts
@@ -1,7 +1,10 @@
-import { Language, TokenizerConfig } from "./tokenizer/index.js";
-import type { Hooks } from "./methods/hooks.js";
-import type { RadixNode } from "./radix-tree/node.js";
-import type { AVLNode } from "./trees/avl/node.js";
+import { Language, TokenizerConfig } from "../tokenizer/index.js";
+import type { Hooks } from "../methods/hooks.js";
+import type { RadixNode } from "../radix-tree/node.js";
+import type { AVLNode } from "../trees/avl/node.js";
+
+export * from "./filters.js";
+export * from "./facets.js";
 
 export type TokenScore = [string, number];
 export type Nullable<T> = T | null;
@@ -21,30 +24,6 @@ export type SearchProperties<
     : TKey
   : never;
 
-export type FacetSorting = "asc" | "desc" | "ASC" | "DESC";
-
-type FacetTypeInterfaces = {
-  string: {
-    limit?: number;
-    offset?: number;
-    sort?: FacetSorting;
-  };
-  number: {
-    ranges: {from: number, to: number}[]
-  };
-  boolean: {
-    true?: boolean;
-    false?: boolean;
-  };
-}
-
-export type FacetsSearch<S extends PropertiesSchema, P extends string = "", K extends keyof S = keyof S> = K extends string
-  ? S[K] extends PropertiesSchema
-    ? FacetsSearch<S[K], `${P}${K}.`>
-    : S[K] extends PropertyType
-      ? { [key in `${P}${K}`]?: FacetTypeInterfaces[S[K]] }
-      : never
-  : never;
 
 export type PropertyType = "string" | "number" | "boolean";
 
@@ -119,6 +98,8 @@ export type BM25Params = {
   d: number;
 };
 
+export type TokenMap = Record<string, TokenScore[]>;
+
 type ResolveTypes<TType> = TType extends "string"
   ? string
   : TType extends "boolean"
@@ -130,8 +111,6 @@ type ResolveTypes<TType> = TType extends "string"
   : never;
 
 type Index = Record<string, RadixNode | AVLNode<number, string[]>>;
-
-export type TokenMap = Record<string, TokenScore[]>;
 
 type FrequencyMap = {
   [property: string]: {
